+++ crates/icu/src/ic/mod.rs
@@ -13,6 +13,8 @@ pub mod icrc {
     pub use icrc_ledger_types::*;
 }
 
-pub use ic_cdk_timers::*;
+pub mod timers {
+    pub use ic_cdk_timers::*;
+}
 
 pub mod structures;