+++ crates/test/src/lib.rs
@@ -1,4 +1,3 @@
-use candid::Principal;
 use icu::prelude::*;
 
 pub static TEST: &str = "test";
@@ -7,10 +6,11 @@ pub static TEST: &str = "test";
 // ICU
 //
 
-icu_start!(TEST);
+icu_start_root!(TEST);
 
-fn _init() {}
-
-async fn _init_async(owner_id: Principal) {
-    log!(Log::Warn, "{owner_id:?}");
+#[update]
+async fn init_async() {
+    ::icu::log!(::icu::Log::Warn, "hello from init_async!!");
 }
+
+export_candid!();